all: add full implementation of instrumented Client
OcRedis is a full drop-in replacement for redis.Redis.
It wraps redis.Redis and uses OpenCensus to collect
traces and metrics for each invocation.
It captures and counts errors, method invocations and delivers
them as traces as well as views that are then exported as metrics.

Sample usage:
```python
import ocredis

def main():
   client = ocredis.OcRedis(host='localhost')

   # Turn on stats collection and reporting.
   # Please remember to then use any of the OpenCensus exporters
   # defined at https://opencensus.io/exporters/supported-exporters/python/
   ocredis.register_views()
   # Enable your exporter of choice here.

   # Now use the client normally just like you would use redis.Redis
   value = client.get('name_here')
   client.del('name_here')
```